+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 ### Added
 - Add local Protobuf conversion instructions to README ([#362](https://github.com/opensearch-project/opensearch-protobufs/pull/362))
+- Add in-place rename support for oneOf fields in post-processing ([#363](https://github.com/opensearch-project/opensearch-protobufs/pull/363))
 - Use official openapi-generator tool ([#364](https://github.com/opensearch-project/opensearch-protobufs/pull/364))
 
 ### Changed